Added support for more projects
- Due to a bug on IMX D1sks are currently disabled, altho implemented
- Created another mat_view
- Various UI improvements

+++ site/backend/src/controller/mints_controller.rs
@@ -6,7 +6,7 @@ use sqlx::{Pool, Postgres};
 #[derive(Deserialize)]
 pub struct Params {
     token_address: String,
-    page: i16,
+    page: i32,
 }
 
 #[get("/api/mint/mints")]
